actualizacion-datos/requirements.txt
requests==2.32.3
boto3==1.35.0
python-dotenv==1.0.1
